Переделал контороллер авторизации
